+++ hspec-src/Flesh/Language/Parser/TestUtil.hs
@@ -33,6 +33,7 @@ import Data.Foldable (for_)
 import Data.Map.Strict (insert, singleton)
 import Data.Text (pack)
 import qualified Flesh.Language.Alias as Alias
+import Flesh.Language.Parser.Alias
 import Flesh.Language.Parser.Char
 import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Error
