+++ src/Flesh/Language/Parser/Capture.hs
@@ -36,6 +36,7 @@ import Control.Monad (MonadPlus, mplus, mzero)
 import Control.Monad.Except (MonadError, catchError, throwError)
 import Control.Monad.Trans.Class (MonadTrans, lift)
 import Control.Monad.Writer.Strict (WriterT, censor, pass, runWriterT, tell)
+import Flesh.Language.Parser.Alias
 import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Input
 import Flesh.Source.Position
