+++ src/Flesh/Language/Pretty.hs
@@ -37,6 +37,7 @@ import Control.Monad.State.Strict (MonadState, StateT, get, put, runStateT)
 import Control.Monad.Trans.Class (lift)
 import Data.Foldable (for_)
 import Data.Map.Lazy (empty)
+import Flesh.Language.Parser.Alias
 import Flesh.Language.Parser.Char
 import Flesh.Language.Parser.Class
 import Flesh.Language.Parser.Error
